requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
apache-beam==2.52.0
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
cloudpickle==2.2.1
crcmod==1.7
datasets==2.15.0
dill==0.3.7
dnspython==2.4.2
docopt==0.6.2
fastavro==1.9.0
fasteners==0.19
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.10.0
grpcio==1.59.3
hdfs==2.7.3
httplib2==0.22.0
huggingface-hub==0.19.4
idna==3.6
Jinja2==3.1.2
Js2Py==0.74
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mwparserfromhell==0.6.5
networkx==3.2.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
objsize==0.6.1
orjson==3.9.10
packaging==23.2
pandas==2.1.3
proto-plus==1.22.3
protobuf==4.25.1
pyarrow==11.0.0
pyarrow-hotfix==0.6
pydot==1.4.2
pyjsparser==2.7.1
pymongo==4.6.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
regex==2023.10.3
requests==2.31.0
rpds-py==0.13.1
six==1.16.0
sympy==1.12
torch==2.1.1
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
xxhash==3.4.1
yarl==1.9.3
zstandard==0.22.0
